flask-api-tutorial-3/web/requirements.txt
Flask
flask_restful
pymongo
numpy
requests
tensorflow
tensorflow_hub
Pillow
